# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise AI Chatbot System API
  description: |
    # Enterprise-Scale Conversational AI Platform

    A production-grade API for conversational AI with enterprise features including:

    - **Multi-Model Support**: GPT-4, Claude-3, and more
    - **Real-time Streaming**: WebSocket and Server-Sent Events
    - **Enterprise SSO**: SAML 2.0, OAuth 2.0, API keys
    - **Advanced Analytics**: Usage tracking, quality scoring
    - **Safety & Compliance**: Content filtering, audit trails
    - **Cost Optimization**: Intelligent model routing, caching
    - **Multi-Region**: Global deployment with failover

    ## Authentication

    This API supports multiple authentication methods:

    - **Bearer Token**: OAuth 2.0 JWT tokens for user authentication
    - **API Key**: Server-to-server authentication with scoped permissions
    - **Session Cookie**: Web application session management

    ## Rate Limiting

    API endpoints are rate limited based on your subscription tier:

    | Tier | Requests/Hour | Concurrent WebSocket |
    |------|---------------|---------------------|
    | Free | 1,000 | 1 |
    | Pro | 10,000 | 10 |
    | Enterprise | 100,000+ | 1,000+ |

    ## Error Handling

    The API uses standard HTTP status codes and returns detailed error information:

    ```json
    {
      "error": {
        "code": "RATE_LIMIT_EXCEEDED",
        "message": "Too many requests",
        "details": {
          "limit": 1000,
...
        }
      }
...

...

...

...

...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  - SessionAuth: []
paths:
...
  /chat/sessions/{sessionId}/messages:
    get:
      tags:
        - Messages
      summary: Get session messages
      description: Retrieve paginated message history for a session
      operationId: getMessages
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: before
          in: query
          schema:
            type: string
            format: date-time
          description: Get messages before this timestamp
        - name: after
          in: query
          schema:
            type: string
            format: date-time
          description: Get messages after this timestamp
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'
    post:
      tags:
        - Messages
      summary: Send message
      description: Send a message and get AI response
      operationId: sendMessage
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
            examples:
              simple_message:
                summary: Simple text message
                value:
                  content: "Hello, how are you?"
                  stream: false
              streaming_message:
                summary: Message with streaming response
                value:
                  content: "Explain quantum computing"
                  stream: true
              message_with_attachments:
                summary: Message with file attachments
                value:
                  content: "Analyze this image"
                  attachments:
                    - type: "image"
                      url: "https://example.com/image.jpg"
                      metadata:
                        filename: "diagram.jpg"
                        size: 1024000
      responses:
        '200':
          description: Message sent successfully (non-streaming)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '202':
          description: Message accepted for streaming
          content:
            application/json:
              schema:
                type: object
                properties:
                  message_id:
                    type: string
                    format: uuid
                  stream_url:
                    type: string
                    format: uri
                    description: WebSocket URL for streaming response
        '400':
          $ref: '#/components/responses/BadRequestError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for server-to-server authentication
    SessionAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: Session cookie for web applications
  parameters:
    SessionIdParam:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Chat session identifier
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of items to skip
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    SendMessageRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          maxLength: 32000
          description: Message content
        stream:
          type: boolean
          default: false
          description: Whether to stream the response
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
          description: File attachments
        context:
          type: object
          description: Additional context for the AI
          properties:
            user_preferences:
              type: object
            conversation_style:
              type: string
              enum:
                - formal
                - casual
                - technical
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - user
            - assistant
            - system
        content:
          type: string
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        metadata:
          type: object
          properties:
            model:
              type: string
            tokens:
              type: integer
            cost:
              type: number
              format: float
            response_time:
              type: number
              format: float
            quality_score:
              type: number
              format: float
        created_at:
          type: string
          format: date-time
    MessageResponse:
      type: object
      properties:
        message:
          $ref: '#/components/schemas/Message'
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
            cost:
              type: number
              format: float
    Attachment:
      type: object
      required:
        - type
        - url
      properties:
        type:
          type: string
          enum:
            - image
            - document
            - audio
            - video
        url:
          type: string
          format: uri
        filename:
          type: string
        size:
          type: integer
          description: File size in bytes
        metadata:
          type: object
          additionalProperties: true
...
    PaginationInfo:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
        next_url:
          type: string
          format: uri
          nullable: true
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
              additionalProperties: true
            request_id:
              type: string
              format: uuid
              description: Request ID for debugging

# Tags for grouping endpoints
...